fixed errors in notifications-arcjet

+++ backend/controllers/notification.controller.js
@@ -6,15 +6,20 @@ import errorHandler from "../middlewares/errorHandler.js";
 
 export const getNotifications = asyncHandler(async (req, res) => {
   const { userId } = getAuth(req);
-  const notifications = Notification.find({
-    to: userId,
+
+  const user = await User.findOne({ clerkId: userId });
+  if (!user) return errorHandler(res, "User Not Found", 404);
+
+  const notifications = await Notification.find({
+    to: user._id,
   })
     .sort({ createdAt: -1 })
     .populate("from", "username firstName lastName profilePicture")
     .populate("post", "content image")
     .populate("comment", "content");
 
-  if (!notifications) return errorHandler(res, "No Notifications Found", 404);
+  if (notifications.length === 0)
+    return errorHandler(res, "No Notifications Found", 404);
   return responseHandler(
     res,
     200,
@@ -26,7 +31,8 @@ export const deleteNotification = asyncHandler(async (req, res) => {
   const { notificationId } = req.params;
   const { userId } = getAuth(req);
 
-  const user = Notification.findOne({ clerkId: userId });
+  const user = User.findOne({ clerkId: userId });
+  if (!user) return errorHandler(res, "User Not Found", 404);
   const notification = Notification.findByIdAndDelete({
     to: user._id,
     _id: notificationId,
